api/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.29.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Automat==22.10.0
backoff==2.2.1
beautifulsoup4==4.12.3
boto3==1.34.48
botocore==1.34.48
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.8.0
cohere==5.3.3
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
dataclasses-json==0.6.4
deepdiff==7.0.1
defusedxml==0.7.1
deptry==0.16.1
distro==1.9.0
emoji==2.12.1
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.109.2
fastavro==1.9.4
filelock==3.13.4
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.1
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
langchain==0.2.6
langchain-anthropic==0.1.15
langchain-cohere==0.1.4
langchain-community==0.2.6
langchain-core==0.2.10
langchain-experimental==0.0.57
langchain-google-cloud-sql-pg==0.4.0
langchain-google-genai==1.0.3
langchain-openai==0.1.1
langchain-text-splitters==0.2.2
langchainhub==0.1.15
langdetect==1.0.9
langsmith==0.1.82
lxml==5.2.2
marshmallow==3.21.1
marshmallow-enum==1.5.1
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
openai==1.30.1
openapi-schema-pydantic==1.2.4
ordered-set==4.1.0
orjson==3.10.0
packaging==23.2
pandas==2.2.2
parsel==1.9.1
pgvector==0.2.5
pluggy==1.5.0
portalocker==2.8.2
Protego==0.3.1
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyDispatcher==2.0.7
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.2.0
pytest==8.2.0
pytest-env==1.1.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.9.0
queuelib==1.7.0
rapidfuzz==3.9.3
regex==2023.12.25
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
spider-client==0.0.48
SQLAlchemy==2.0.29
sse-starlette==2.1.0
starlette==0.36.3
tabulate==0.9.0
tavily-python==0.3.3
tenacity==8.2.3
tiktoken==0.6.0
tldextract==5.1.2
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.2
Twisted==24.3.0
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
unstructured==0.14.9
unstructured-client==0.23.7
upstash-redis==1.0.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.29.0
w3lib==2.2.1
wrapt==1.16.0
yarl==1.9.4
zope.interface==6.4.post2
